refactor(KERN): remove unused summary overview components
