+++ app/components/content/summaryOverview/SummaryOverviewBoxItem.tsx
@@ -1,26 +1,20 @@
 import type { UserData } from "~/domains/userData";
-import { isGeldEinklagenLongTextField } from "~/domains/geldEinklagen/formular/klage-erstellen/longTextFieldConfig";
 import type { FieldItems, SummaryOverviewBoxItemType } from "./types";
 import { type Translations } from "~/services/translations/getTranslationByKey";
+import { translations as staticTranslations } from "~/services/translations/translations";
 import {
   extractFieldItemsFromInlineItems,
   getItemValueBox,
 } from "./getItemValueBox";
+import { hasMoneyValidationSchema, hasNonEmptyLongTextField } from "./helper";
 
 type Props = SummaryOverviewBoxItemType & {
   readonly userData: UserData;
   readonly translations: Translations;
+  readonly pathname: string;
 };
 
 const SCROLLABLE_BOX_ROWS = 10;
-
-const hasNonEmptyLongTextField = (fieldItems: FieldItems) =>
-  fieldItems.some(
-    ({ fieldName, fieldValue }) =>
-      isGeldEinklagenLongTextField(fieldName) &&
-      typeof fieldValue === "string" &&
-      fieldValue.trim().length > 0,
-  );
 
 const SummaryValueOverflowContainer = ({
   children,
@@ -36,11 +30,13 @@ const SummaryValueOverflowContainer = ({
   }
 
   return (
-    <div
-      className="border-blue-500 p-[16px] mt-8 w-full overflow-auto resize-y box-border border-1 whitespace-pre-wrap"
-      style={{ blockSize: `${SCROLLABLE_BOX_ROWS * 1.5}rem` }}
-    >
-      {children}
+    <div className="kern-form-input">
+      <div
+        className="kern-form-input__input pt-kern-space-small! overflow-auto resize-y box-border whitespace-pre-wrap"
+        style={{ blockSize: `${SCROLLABLE_BOX_ROWS * 1.5}rem` }}
+      >
+        {children}
+      </div>
     </div>
   );
 };
@@ -50,26 +46,37 @@ const SummaryOverviewBoxItem = ({
   translations,
   title,
   inlineItems,
+  pathname,
 }: Props) => {
-  const itemValue = getItemValueBox(translations, userData, inlineItems);
-  if (itemValue.trim() === "") return null;
+  const rawItemValue = getItemValueBox(translations, userData, inlineItems);
+  if (rawItemValue.trim() === "") return null;
 
   const fieldItems = extractFieldItemsFromInlineItems(userData, inlineItems);
+  const shouldAppendEuroWord = hasMoneyValidationSchema(pathname, fieldItems);
+  const itemValue = shouldAppendEuroWord
+    ? `${rawItemValue} ${staticTranslations.currency.euro.de}`
+    : rawItemValue;
 
   return (
-    <div>
+    <>
       {title && (
-        <dt data-testid="summary-box-item-title" className="ds-label-01-bold">
+        <dt
+          data-testid="summary-box-item-title"
+          className="kern-label text-kern-static-medium w-full"
+        >
           {title}
         </dt>
       )}
 
-      <dd data-testid="summary-box-item-value">
+      <dd
+        data-testid="summary-box-item-value"
+        className="kern-body text-kern-static-medium w-full h-full"
+      >
         <SummaryValueOverflowContainer fieldItems={fieldItems}>
           {itemValue}
         </SummaryValueOverflowContainer>
       </dd>
-    </div>
+    </>
   );
 };
