+++ app/components/content/summaryOverview/__test__/SummaryOverviewBox.test.tsx
@@ -2,7 +2,7 @@ import { render } from "@testing-library/react";
 import { describe, it, vi, expect } from "vitest";
 import { type HeadingProps } from "~/components/common/Heading";
 import { useFormFlow } from "~/components/formFlowContext";
-import SummaryOverviewBox from "../SummaryOverviewBox";
+import SummaryOverviewBox from "~/components/content/summaryOverview/SummaryOverviewBox";
 
 vi.mock("~/components/formFlowContext", () => ({
   useFormFlow: vi.fn(),
@@ -51,7 +51,7 @@ describe("SummaryOverviewBox", () => {
   });
 
   it("should render the heading when title is provided", () => {
-    const { getByTestId } = render(
+    const { getByRole } = render(
       <SummaryOverviewBox
         boxId={boxId}
         stepId={stepId}
@@ -61,11 +61,11 @@ describe("SummaryOverviewBox", () => {
       />,
     );
 
-    expect(getByTestId("heading")).toHaveTextContent(titleMock.text ?? "");
+    expect(getByRole("heading")).toHaveTextContent(titleMock.text ?? "");
   });
 
   it("should render the heading when title is provided with arrayPositionTitle", () => {
-    const { getByTestId } = render(
+    const { getByRole } = render(
       <SummaryOverviewBox
         boxId={boxId}
         stepId={stepId}
@@ -76,7 +76,7 @@ describe("SummaryOverviewBox", () => {
       />,
     );
 
-    expect(getByTestId("heading")).toHaveTextContent(`${titleMock.text} 2`);
+    expect(getByRole("heading")).toHaveTextContent(`${titleMock.text} 2`);
   });
 
   it("should render SummaryOverviewBoxItem for each field", () => {
