+++ app/components/content/summaryOverview/helper.ts
