--- app/components/kern/summaryOverview/__test__/getItemValueBox.test.ts
@@ -1,170 +1,0 @@
-import { describe, expect, test } from "vitest";
-import { type UserData } from "~/domains/userData";
-import { type Translations } from "~/services/translations/getTranslationByKey";
-import {
-  getItemValueBox,
-  extractFieldItemsFromInlineItems,
-} from "../getItemValueBox";
-
-describe("getItemValueBox", () => {
-  test("returns translated value if a direct translation exists", () => {
-    const translations: Translations = {
-      "status.active": "Aktiv",
-    };
-    const userData: UserData = { status: "active" };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status" },
-    ]);
-    expect(actual).toBe("Aktiv");
-  });
-
-  test("returns empty value when item value is empty", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: "" };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status" },
-    ]);
-
-    expect(actual).toBe("");
-  });
-
-  test("returns field.value translation if no direct translation exists", () => {
-    const translations: Translations = {
-      "status.value": "Default Status",
-    };
-    const userData: UserData = { status: "unknown" };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status" },
-    ]);
-
-    expect(actual).toBe("Default Status");
-  });
-
-  test("returns the original item value if no translations exist", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: "pending" };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status" },
-    ]);
-
-    expect(actual).toBe("pending");
-  });
-
-  test("returns the original item value when translation.value is not available", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: "approved" };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status" },
-    ]);
-
-    expect(actual).toBe("approved");
-  });
-
-  test("returns value when it is nested in an object", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: { approved: "true" } };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status.approved" },
-    ]);
-
-    expect(actual).toBe("true");
-  });
-
-  test("returns empty value when it does not exist in the object", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: { approved: "true" } };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status.fail" },
-    ]);
-
-    expect(actual).toBe("");
-  });
-
-  test("returns concatenated values for multiple fields", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: "active", role: "admin" };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status" },
-      { field: "role" },
-    ]);
-
-    expect(actual).toBe("active admin");
-  });
-
-  test("returns emptyValuePlaceholder when item value is empty and placeholder is provided", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: "" };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status", emptyValuePlaceholder: "No Status" },
-    ]);
-
-    expect(actual).toBe("No Status");
-  });
-
-  test("returns emptyValuePlaceholder when nested item value is empty and placeholder is provided", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: { approved: "" } };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status.approved", emptyValuePlaceholder: "Not Approved" },
-    ]);
-
-    expect(actual).toBe("Not Approved");
-  });
-
-  test("returns concatenated values with placeholders for multiple fields", () => {
-    const translations: Translations = {};
-    const userData: UserData = { status: "", role: "admin" };
-
-    const actual = getItemValueBox(translations, userData, [
-      { field: "status", emptyValuePlaceholder: "No Status" },
-      { field: "role" },
-    ]);
-
-    expect(actual).toBe("No Status admin");
-  });
-});
-
-describe("extractFieldItemsFromInlineItems", () => {
-  it("extracts fieldName and fieldValue pairs for simple and nested fields", () => {
-    const userData: UserData = {
-      sachverhaltBegruendung: "Free text",
-      weiterePersonen: { buchungsnummer: "ABCDEF10" },
-    };
-
-    const actual = extractFieldItemsFromInlineItems(userData, [
-      { field: "sachverhaltBegruendung" },
-      { field: "weiterePersonen.buchungsnummer" },
-    ]);
-
-    expect(actual).toEqual([
-      { fieldName: "sachverhaltBegruendung", fieldValue: "Free text" },
-      {
-        fieldName: "weiterePersonen.buchungsnummer",
-        fieldValue: "ABCDEF10",
-      },
-    ]);
-  });
-
-  it("uses only the extract field name for array-prefixed field", () => {
-    const userData: UserData = {
-      weiterePersonen: [{ vorname: "Mina" }],
-      vorname: "Erika",
-    };
-
-    const actual = extractFieldItemsFromInlineItems(userData, [
-      { field: `weiterePersonen#vorname` },
-    ]);
-
-    expect(actual).toEqual([{ fieldName: "vorname", fieldValue: "Erika" }]);
-  });
-});
