--- app/components/kern/summaryOverview/addArrayIndexToPathUrl.ts
@@ -1,13 +1,0 @@
-export const addArrayIndexToPathUrl = (path: string, index: number): string => {
-  const segments = path.split("/");
-
-  if (segments.length < 2) {
-    return path;
-  }
-
-  const startSegment = segments.length - 1;
-
-  segments.splice(startSegment, 0, index.toString());
-
-  return segments.join("/");
-};
